Rendering props for count and header using the handleForm function from app.js
